refactor: consolidate cardano-serialization-lib package into core
