--- packages/cardano-serialization-lib/README.md
@@ -1,5 +1,0 @@
-# Cardano JS SDK | cardano-serialization-lib
-
-This package contains the integration with [cardano-serialization-lib].
-
-[cardano-serialization-lib]: https://github.com/Emurgo/cardano-serialization-lib
