--- packages/cardano-serialization-lib/src/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  "extends": ["../../../.eslintrc.js"],
-  "parserOptions": {
-    "project": "./tsconfig.json",
-    "tsconfigRootDir": __dirname
-  }
-}
