--- packages/cardano-serialization-lib/src/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../dist/",
-    "rootDir": "."
-  }
-}
