--- packages/cardano-serialization-lib/test/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  "extends": ["../../../test/.eslintrc.js"],
-  "parserOptions": {
-    "project": "./tsconfig.json",
-    "tsconfigRootDir": __dirname
-  }
-}
