--- packages/cardano-serialization-lib/test/jest.config.js
@@ -1,11 +1,0 @@
-const { pathsToModuleNameMapper } = require('ts-jest/utils')
-const { compilerOptions } = require('./tsconfig')
-
-module.exports = {
-  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths),
-  preset: 'ts-jest',
-  transform: {
-    "^.+\\.test.ts?$": "ts-jest"
-  },
-  testTimeout: 120000
-}
