--- packages/cardano-serialization-lib/test/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@src/*": ["../src/*"]
-    }
-  },
-  "references": [{ "path": "../src" }]
-}
