+++ packages/cip2/jest.config.js
@@ -4,5 +4,6 @@ module.exports = {
   transform: {
     "^.+\\.test.ts?$": "ts-jest"
   },
+  coveragePathIgnorePatterns: ['\.config\.js'],
   testTimeout: process.env.CI ? 120000 : 12000,
 }
