+++ packages/cip2/src/tsconfig.json
@@ -5,7 +5,6 @@
     "rootDir": "."
   },
   "references": [
-    { "path": "../../cardano-serialization-lib/src" },
     { "path": "../../core/src" }
   ]
 }
