+++ packages/cip2/src/types.ts
@@ -1,4 +1,4 @@
-import { CSL } from '@cardano-sdk/cardano-serialization-lib';
+import { CSL } from '@cardano-sdk/core';
 
 export interface SelectionResult {
   selection: {
