+++ packages/core/package.json
@@ -18,11 +18,12 @@
     "coverage": "shx echo No coverage report for this package"
   },
   "devDependencies": {
-    "@cardano-sdk/cardano-serialization-lib": "0.1.0",
     "shx": "^0.3.3"
   },
   "dependencies": {
     "@cardano-ogmios/schema": "^4.0.0-beta.6",
+    "@emurgo/cardano-serialization-lib-nodejs": "8.0.0",
+    "@emurgo/cardano-serialization-lib-browser": "^8.0.0",
     "buffer": "^6.0.3"
   }
 }
