+++ packages/core/src/index.ts
@@ -4,4 +4,5 @@ export * as Ogmios from './Ogmios';
 export * from './Genesis';
 export * from './Provider';
 export * as Transaction from './Transaction';
+export * from './CSL';
 export * from './util';
