+++ packages/core/src/tsconfig.json
@@ -3,10 +3,5 @@
   "compilerOptions": {
     "outDir": "../dist/",
     "rootDir": "."
-  },
-  "references": [
-    {
-      "path": "../../cardano-serialization-lib/src"
-    }
-  ]
+  }
 }
