+++ packages/wallet/package.json
@@ -28,7 +28,6 @@
   },
   "dependencies": {
     "@cardano-ogmios/schema": "^4.0.0-beta.6",
-    "@cardano-sdk/cardano-serialization-lib": "0.1.0",
     "@cardano-sdk/cip2": "0.1.0",
     "@cardano-sdk/core": "0.1.0",
     "ts-log": "^2.2.3"
