+++ packages/wallet/src/tsconfig.json
@@ -5,7 +5,6 @@
     "rootDir": "."
   },
   "references": [
-    { "path": "../../cardano-serialization-lib/src" },
     { "path": "../../cip2/src" },
     { "path": "../../core/src" }
   ]
