+++ packages/wallet/test/createTransactionInternals.test.ts
@@ -1,7 +1,13 @@
-import { loadCardanoSerializationLib, CardanoSerializationLib, CSL } from '@cardano-sdk/cardano-serialization-lib';
 import { createTransactionInternals } from '@src/createTransactionInternals';
 import { InputSelector, roundRobinRandomImprove } from '@cardano-sdk/cip2';
-import { Cardano, CardanoProvider, Ogmios } from '@cardano-sdk/core';
+import {
+  loadCardanoSerializationLib,
+  CardanoSerializationLib,
+  CSL,
+  Cardano,
+  CardanoProvider,
+  Ogmios
+} from '@cardano-sdk/core';
 import { providerStub } from './ProviderStub';
 import { UtxoRepository } from '@src/UtxoRepository';
 import { InMemoryUtxoRepository } from '@src/InMemoryUtxoRepository';
