+++ scripts/pack.sh
@@ -4,7 +4,6 @@ set -euo pipefail
 
 npm pack --cwd ./packages/blockfrost && \
 npm pack --cwd ./packages/cardano-graphql-db-sync && \
-npm pack --cwd ./packages/cardano-serialization-lib && \
 npm pack --cwd ./packages/cip2 && \
 npm pack --cwd ./packages/cip30 && \
 npm pack --cwd ./packages/core && \
