+++ scripts/publish.sh
@@ -4,7 +4,6 @@ set -euo pipefail
 
 npm publish --cwd ./packages/blockfrost && \
 npm publish --cwd ./packages/cardano-graphql-db-sync && \
-npm publish --cwd ./packages/cardano-serialization-lib && \
 npm publish --cwd ./packages/cip2 && \
 npm publish --cwd ./packages/cip30 && \
 npm publish --cwd ./packages/core && \
